[REF] l10n_br_fiscal: document.mixin.fields->document.mixin
it's a simple mv followed by the insertion of _inherit = "l10n_br_fiscal.document.mixin.methods"
and a correction: _description = "Document Fiscal Mixin"
